Represent Redis integers as Long instead of Int
Redis integers seem to be 64-bit, making Long a better representation.

+++ src/main/scala/org/programmiersportgruppe/redis/UnifiedProtocolLexers.scala
@@ -10,7 +10,7 @@ case object ErrorSigil extends Sigil('-')
 case object IntegerSigil extends Sigil(':')
 case object BulkSigil extends Sigil('$')
 case object MultiBulkSigil extends Sigil('*')
-case class IntegerNewlineToken(value: Int) extends Token
+case class IntegerNewlineToken(value: Long) extends Token
 case class ByteStringToken(value: ByteString) extends Token
 case object NewlineToken extends Token
